refactor(git_ops)!: extract temp worktree and GitHub logic into modules
### What

- Move `TempWorktree`, `is_in_temp_worktree`, and `cleanup_old_patches` from `git_ops.rs` into a new `git_temp_worktree` module
- Extract GitHub issue listing logic from `git_ops.rs` into a new `github_ops` module (`github_list_issues`)
- Remove corresponding functions and imports from `git_ops.rs`
- Update `main.rs` and `lib.rs` to import and re-export the new modules
- Adjust tests to import `github_ops` and call `github_list_issues`, and update assertion style

### Why

- Enforce separation of concerns: isolate git worktree management and GitHub operations
- Simplify the core `git_ops` module and improve code organization and maintainability

### Bigger Picture

- Modularization sets the stage for further extensions (e.g., support for different remote providers)
- Improves testability by encapsulating functionality in focused modules
- TODO: Add module-level documentation and update any downstream consumers

### Impact

- Breaking change: `git_list_issues` has been removed; consumers must migrate to `github_list_issues`
