+++ src/git_ops.rs
@@ -1,244 +1,13 @@
 use crate::tui::{render_message, App};
-use once_cell::sync::OnceCell;
 use ratatui::style::Color;
 use ratatui::{backend::Backend, Terminal};
 use std::error::Error;
-use std::io::Write;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::process::Command;
-use std::sync::Mutex;
 
 const AUTOCOMMIT_BRANCH_NAME: &str = "gh-autopr-index-autocommit";
 const AUTOSTASH_NAME: &str = "gh-autopr-index-autostash";
 const MAX_DIFF_BYTES: usize = 200 * 1024; // 200 KiB
-
-/// RAII guard for the temp worktree
-pub struct TempWorktree {
-    path: PathBuf,
-    orig_root: PathBuf,
-    orig_branch: String,
-}
-
-impl TempWorktree {
-    /// Enter a detached worktree that lives in `.git/autopr-wt-<uuid>`.
-    pub fn enter() -> Result<Self, Box<dyn Error>> {
-        // 1. capture original location + branch + changes ------------------------
-        let orig_root = PathBuf::from(
-            String::from_utf8(
-                Command::new("git")
-                    .args(["rev-parse", "--show-toplevel"])
-                    .output()?
-                    .stdout,
-            )?
-            .trim(),
-        );
-        let orig_branch = String::from_utf8(
-            Command::new("git")
-                .args(["rev-parse", "--abbrev-ref", "HEAD"])
-                .output()?
-                .stdout,
-        )?
-        .trim()
-        .to_owned();
-
-        let staged_patch = Command::new("git")
-            .args(["diff", "--staged", "--binary"])
-            .output()?
-            .stdout;
-        let unstaged_patch = Command::new("git")
-            .args(["diff", "--binary"])
-            .output()?
-            .stdout;
-        let untracked_list = String::from_utf8(
-            Command::new("git")
-                .args(["ls-files", "--others", "--exclude-standard", "-z"])
-                .output()?
-                .stdout,
-        )?;
-
-        // 2. construct a path on the *same* filesystem (inside .git) -------------
-        let git_dir = PathBuf::from(
-            String::from_utf8(
-                Command::new("git")
-                    .args(["rev-parse", "--git-dir"])
-                    .output()?
-                    .stdout,
-            )?
-            .trim(),
-        );
-        let ts = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)?
-            .as_millis();
-        let path = git_dir.join(format!("autopr-wt-{}", ts));
-
-        // 3. create detached worktree at HEAD ------------------------------------
-        let out = Command::new("git")
-            .args(["worktree", "add", "--detach", path.to_str().unwrap()])
-            .output()?;
-        if !out.status.success() {
-            return Err(format!(
-                "git worktree add failed: {}",
-                String::from_utf8_lossy(&out.stderr)
-            )
-            .into());
-        }
-
-        // 4. hop into that directory --------------------------------------------
-        std::env::set_current_dir(&path)?;
-
-        // 5. make the worktree point at the **same branch name** the user had.
-        //    Use --force so this works even if that branch is already active
-        //    in another worktree.
-        //
-        //    ① direct switch (branch exists locally)          ───────────────────
-        //    ② otherwise try to track the remote branch       ───────────────────
-        //    ③ as last resort create an *orphan* local branch ───────────────────
-        let mut ok = Command::new("git")
-            .args([
-                "switch",
-                "--force",
-                "--ignore-other-worktrees",
-                &orig_branch,
-            ]) // <-- --force here
-            .status()?
-            .success();
-
-        if !ok {
-            // remote may exist – create local branch that tracks it
-            ok = Command::new("git")
-                .args([
-                    "switch",
-                    "--force", // allow switching to a branch that already exists
-                    "--ignore-other-worktrees",
-                    "-c",
-                    &orig_branch,
-                    "--track",
-                    &format!("origin/{}", orig_branch),
-                ])
-                .status()?
-                .success();
-        }
-
-        if !ok {
-            // last resort: create orphaned branch with same name
-            let out = Command::new("git")
-                .args([
-                    "switch",
-                    "--force",
-                    "--ignore-other-worktrees",
-                    "-c",
-                    &orig_branch,
-                ])
-                .output()?;
-            if !out.status.success() {
-                return Err(format!(
-                    "failed to switch to temp branch {orig_branch}: {}",
-                    String::from_utf8_lossy(&out.stderr)
-                )
-                .into());
-            }
-        }
-
-        // ── 6. replay the dirty state inside the temp work-tree ────────────────
-
-        // 6a. staged patch → index only
-        if !staged_patch.is_empty() {
-            let mut child = Command::new("git")
-                .args(["apply", "--cached", "-"])
-                .stdin(std::process::Stdio::piped())
-                .spawn()?;
-            child.stdin.as_mut().unwrap().write_all(&staged_patch)?;
-            if !child.wait()?.success() {
-                return Err("failed to apply staged patch".into());
-            }
-        }
-
-        // 6b. unstaged patch → working tree
-        if !unstaged_patch.is_empty() {
-            let mut child = Command::new("git")
-                .args(["apply", "-"])
-                .stdin(std::process::Stdio::piped())
-                .spawn()?;
-            child.stdin.as_mut().unwrap().write_all(&unstaged_patch)?;
-            if !child.wait()?.success() {
-                return Err("failed to apply unstaged patch".into());
-            }
-        }
-
-        // 6c. untracked files
-        if !untracked_list.is_empty() {
-            for path in untracked_list.split('\0').filter(|p| !p.is_empty()) {
-                let from = orig_root.join(path);
-                let to = path; // relative inside temp WT
-
-                // Check if source file exists before trying to copy
-                // (files might have been renamed/moved and no longer exist at original path)
-                if !from.exists() {
-                    eprintln!(
-                        "Warning: Skipping untracked file copy - source doesn't exist: {}",
-                        from.display()
-                    );
-                    continue;
-                }
-
-                if let Some(parent) = Path::new(to).parent() {
-                    if let Err(e) = fs_err::create_dir_all(parent) {
-                        eprintln!(
-                            "Warning: Failed to create directory {}: {}",
-                            parent.display(),
-                            e
-                        );
-                        continue;
-                    }
-                }
-
-                if let Err(e) = fs_err::copy(&from, to) {
-                    eprintln!(
-                        "Warning: Failed to copy untracked file from {} to {}: {}",
-                        from.display(),
-                        to,
-                        e
-                    );
-                    continue;
-                }
-            }
-        }
-
-        // All done – temp work-tree now has *exact* dirty state.
-        Ok(Self {
-            path,
-            orig_root,
-            orig_branch,
-        })
-    }
-}
-
-impl Drop for TempWorktree {
-    fn drop(&mut self) {
-        // 1. hop back to the original worktree (ignore errors to stay unwind-safe)
-        let _ = std::env::set_current_dir(&self.orig_root);
-
-        // 2. if we’re detached, return to the original branch
-        let is_detached = Command::new("git")
-            .args(["symbolic-ref", "--quiet", "HEAD"])
-            .status()
-            .map(|s| !s.success())
-            .unwrap_or(false);
-
-        if is_detached {
-            let _ = Command::new("git")
-                .args(["switch", &self.orig_branch])
-                .status();
-        }
-
-        // 3. remove the temporary worktree
-        let _ = Command::new("git")
-            .args(["worktree", "remove", "--force", self.path.to_str().unwrap()])
-            .status();
-
-        let _ = fs_err::remove_dir_all(&self.path); // belt-and-suspenders
-    }
-}
 
 pub fn git_ensure_in_repo(app: &mut App) -> Result<(), Box<dyn Error>> {
     let output = Command::new("git")
@@ -644,9 +413,6 @@ pub fn git_fetch_main(
     current_branch: &String,
     main_branch: &String,
 ) -> Result<(), Box<dyn Error>> {
-    // Clean up patch files older than 30 days to prevent accumulation
-    let _ = cleanup_old_patches(app, 30);
-
     if current_branch == main_branch {
         let had_staged_changes = git_has_staged_changes()?;
         if had_staged_changes {
@@ -1098,139 +864,6 @@ pub fn create_or_update_pull_request(
             app.add_log("INFO", format!("Pull request URL: {}", url.trim()));
         }
     }
-    Ok(())
-}
-
-// Example GitHub issues JSON output:
-/*
-[
-  {
-    "body": "This is a body of the GH issue.",
-    "labels": [
-      {
-        "id": "LA_kwDOOTdaS88AAAAB9JPIwX",
-        "name": "bug",
-        "description": "Something isn't working",
-        "color": "d73a4a"
-      }
-    ],
-    "number": 42,
-    "title": "This is a title of the GH issue."
-  }
-]
-*/
-pub fn git_list_issues(app: &mut App) -> Result<String, Box<dyn Error>> {
-    // Initialize cache if not already initialized
-    let cache = ISSUES_CACHE.get_or_init(|| Mutex::new(None));
-    let mut cache = cache.lock().unwrap();
-
-    // Return cached data if available
-    if let Some(cached_data) = cache.as_ref() {
-        app.add_log("INFO", "Using cached GitHub issues");
-        return Ok(cached_data.clone());
-    }
-
-    // Cache miss - fetch from GitHub
-    let output = Command::new("gh")
-        .args(["issue", "list", "--json", "number,title,labels,body"])
-        .output()?;
-
-    let stderr = String::from_utf8_lossy(&output.stderr).to_string();
-    if !output.status.success() {
-        // Handle specific error cases
-        if stderr.contains("has disabled issues") {
-            app.add_log("WARN", "Repository has issues disabled");
-            let empty_list = "[]".to_string();
-            *cache = Some(empty_list.clone());
-            return Ok(empty_list);
-        }
-        app.add_error(stderr);
-        return Err("Failed to list issues".into());
-    }
-
-    let json_str = String::from_utf8(output.stdout)?;
-    app.add_log("INFO", "Successfully retrieved fresh GitHub issues");
-
-    // Update cache
-    *cache = Some(json_str.clone());
-
-    Ok(json_str)
-}
-
-static ISSUES_CACHE: OnceCell<Mutex<Option<String>>> = OnceCell::new();
-
-/// Check if we're currently in a temporary worktree
-pub fn is_in_temp_worktree() -> bool {
-    if let Ok(current_dir) = std::env::current_dir() {
-        if let Some(dir_name) = current_dir.file_name().and_then(|n| n.to_str()) {
-            return dir_name.starts_with("autopr-wt-");
-        }
-    }
-    false
-}
-
-/// Clean up old patch files to prevent accumulation
-/// Removes patch files older than `days_old` days
-pub fn cleanup_old_patches(app: &mut App, days_old: u64) -> Result<(), Box<dyn Error>> {
-    let git_dir = PathBuf::from(
-        String::from_utf8(
-            Command::new("git")
-                .args(["rev-parse", "--git-dir"])
-                .output()?
-                .stdout,
-        )?
-        .trim(),
-    );
-    let patches_dir = git_dir.join("gh-autopr-patches");
-
-    if !patches_dir.exists() {
-        return Ok(());
-    }
-
-    let cutoff_time = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)?
-        .as_secs()
-        - (days_old * 24 * 60 * 60);
-
-    let mut cleaned_count = 0;
-    for entry in fs_err::read_dir(&patches_dir)? {
-        let entry = entry?;
-        let path = entry.path();
-
-        if path.is_file() && path.extension().is_some_and(|ext| ext == "patch") {
-            if let Some(filename) = path.file_name().and_then(|n| n.to_str()) {
-                // Extract timestamp from filename (format: staged-TIMESTAMP.patch or unstaged-TIMESTAMP.patch)
-                if let Some(timestamp_str) =
-                    filename.split('-').nth(1).and_then(|s| s.split('.').next())
-                {
-                    if let Ok(timestamp) = timestamp_str.parse::<u64>() {
-                        if timestamp < cutoff_time {
-                            if let Err(e) = fs_err::remove_file(&path) {
-                                app.add_log(
-                                    "WARN",
-                                    format!(
-                                        "Failed to remove old patch file {}: {}",
-                                        path.display(),
-                                        e
-                                    ),
-                                );
-                            } else {
-                                cleaned_count += 1;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    if cleaned_count > 0 {
-        app.add_log(
-            "INFO",
-            format!("Cleaned up {} old patch files", cleaned_count),
-        );
-    }
-
     Ok(())
 }
