+++ src/lib.rs
@@ -1,8 +1,12 @@
 pub mod git_ops;
+pub mod git_temp_worktree;
+pub mod github_ops;
 pub mod gpt_ops;
 pub mod tui;
 
 // Re-export commonly used items
 pub use git_ops::*;
+pub use git_temp_worktree::*;
+pub use github_ops::*;
 pub use gpt_ops::*;
 pub use tui::*;
